KNIME-1704 FingerprintReader FingerprintWriter nodes: migration to KNIME NIO (#137)
Authored-by: Roman Balabanov <roman-1.balabanov_ext@novartis.com>
