+++ org.rdkit.knime.nodes/src-deprecated/org/rdkit/knime/nodes/fingerprintreader/RDKitFingerprintReaderNodeFactory.xml
