+++ org.rdkit.knime.nodes/src-deprecated/org/rdkit/knime/nodes/fingerprintwriter/RDKitFingerprintWriterNodeFactory.xml
