+++ org.rdkit.knime.nodes/src/org/rdkit/knime/nodes/fingerprintreader/RDKitFingerprintReaderV2NodeFactory.xml
@@ -1,0 +1,94 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE knimeNode>
+<knimeNode icon="default.png" type="Source" xmlns="http://knime.org/node/v4.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://knime.org/node/v4.1 http://knime.org/node/v4.1.xsd">
+    <name>RDKit Fingerprint Reader</name>
+    <shortDescription>
+        Node to read fingerprints from a FPS file into an output table.
+    </shortDescription>
+
+    <fullDescription>
+        <intro>
+            This node reads an FPS file with its fingerprint records into a KNIME table.
+            The format of the .fps file is mentioned here: https://jcheminf.springeropen.com/articles/10.1186/1758-2946-5-S1-P36
+        </intro>
+
+        <tab name="Options">
+            <option name="Input File">
+                The location of the FPS file containing fingerprint records.
+                <br/><br/>
+                <b>Read from</b>
+                <br/>
+                Select a file system which stores the model you want to read. There
+                are four default file system options to choose from:
+                <br />
+                <ul>
+                    <li><i>Local File System:</i> Allows you to select a file from your
+                        local system.
+                    </li>
+                    <li><i>Mountpoint:</i> Allows you to read from a mountpoint.
+                        When selected, a new drop-down menu appears to choose the mountpoint.
+                        Unconnected mountpoints are greyed out but can still be selected (note that browsing is disabled in this case).
+                        Go to the KNIME Explorer and connect to the mountpoint to enable browsing.
+                        A mountpoint is displayed in red if it was previously selected but is no longer available.
+                        You won't be able to save the dialog as long as you don't select a valid i.e. known mountpoint.
+                    </li>
+                    <li><i>Relative to:</i> Allows you to choose whether to resolve the path
+                        relative to the current mountpoint, current workflow or the current workflow's data area.
+                        When selected, a new drop-down menu appears to choose which of the two options to use.
+                    </li>
+                    <li><i>Custom/KNIME URL:</i> Allows to specify a URL (e.g. file://, http:// or knime:// protocol).
+                        When selected, a spinner appears that allows you to specify the desired connection and read timeout in milliseconds.
+                        In case it takes longer to connect to the host / read the file, the node fails to execute.
+                        Browsing is disabled for this option.
+                    </li>
+                </ul>
+                It is possible to use other file systems with this node. Therefore,
+                you have to enable the file system connection input port of this node
+                by clicking the
+                <b>...</b>
+                in the bottom left corner of the node's icon and choose
+                <i>Add File System Connection port</i>
+                .
+                <br />
+                Afterwards, you can simply connect the desired connector node to this node.
+                The file system connection will then be shown in the drop-down menu.
+                It is greyed out if the file system is not connected in which case you have to (re)execute the connector node first.
+                Note: The default file systems listed above can't be selected if a file system is provided via the input port.
+                <br/><br/>
+                <b>File/URL</b>
+                <br/>
+                Enter a URL when reading from <i>Custom/KNIME URL</i>, otherwise enter a path to a file. The
+                required syntax of a path depends on the chosen file system, such as &quot;C:\path\to\file&quot; (Local
+                File System on Windows) or  &quot;/path/to/file&quot; (Local
+                File System on Linux/MacOS and Mountpoint). For file systems connected via input port, the node
+                description of the respective connector node describes the required path format.
+                You can also choose a previously selected file from the drop-down list, or select a location
+                from the &quot;Browse...&quot; dialog.
+                Note that browsing is disabled in some cases:
+                <ul>
+                    <li><i>Custom/KNIME URL:</i> Browsing is always disabled.</li>
+                    <li><i>Mountpoint:</i> Browsing is disabled if the selected mountpoint isn't connected.
+                        Go to the KNIME Explorer and connect to the mountpoint to enable browsing.</li>
+                    <li><i>File systems provided via input port:</i>
+                        Browsing is disabled if the connector node hasn't been executed since the workflow has been opened.
+                        (Re)execute the connector node to enable browsing.</li>
+                </ul>
+                <i>The location can be exposed as or automatically set via a
+                </i><a href="https://docs.knime.com/latest/analytics_platform_file_handling_guide/index.html#path">
+                <i>path flow variable.</i></a>
+            </option>
+            <option name="Use IDs from file as row IDs (Requires unique IDs!)">
+                Flag to determine, if IDs read from a fingerprint record of the file shall be used as row IDs.
+                This will fail, if the FPS file does not contain unique fingerprint IDs. </option>
+           </tab>
+    </fullDescription>
+
+    <ports>
+        <dynInPort insert-before="0"
+                   name="File system connection"
+                   group-identifier="File System Connection">
+            The file system connection.
+        </dynInPort>
+        <outPort index="0" name="Fingerprint table">Table containing fingerprints.</outPort>
+    </ports>
+</knimeNode>
