+++ org.rdkit.knime.nodes/src/org/rdkit/knime/nodes/fingerprintwriter/RDKitFingerprintWriterV2NodeFactory.xml
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE knimeNode>
+<knimeNode icon="default.png" type="Sink" xmlns="http://knime.org/node/v4.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://knime.org/node/v4.1 http://knime.org/node/v4.1.xsd">
+    <name>RDKit Fingerprint Writer</name>
+
+    <shortDescription>
+        Node to write fingerprints from an input table to an FPS file.
+    </shortDescription>
+
+    <fullDescription>
+        <intro>
+            This node writes an FPS file using fingerprints from DenseBitVector cells of an input table.
+            If the file exists already it will not be overridden by default.
+            The format of the FPS file is mentioned here: https://jcheminf.springeropen.com/articles/10.1186/1758-2946-5-S1-P36
+        </intro>
+
+        <tab name="Options">
+            <option name="Output File">
+                The location of the output file where the FPS file will be created.
+                <br/><br/>
+                <b>Write to</b>
+                <br/>
+                Select a file system in which you want to store the file. There are four default file system options to
+                choose from:
+                <br />
+                <ul>
+                    <li><i>Local File System:</i> Allows you to select a location in your local system.
+                    </li>
+                    <li><i>Mountpoint:</i> Allows you to write to a mountpoint. When selected, a new drop-down menu appears
+                        to choose the mountpoint. Unconnected mountpoints are greyed out but can still be selected (note that
+                        browsing is disabled in this case). Go to the KNIME Explorer and connect to the mountpoint to enable
+                        browsing. A mountpoint is displayed in red if it was previously selected but is no longer available.
+                        You won't be able to save the dialog as long as you don't select a valid i.e. known mountpoint.
+                    </li>
+                    <li><i>Relative to:</i> Allows you to choose whether to resolve the path relative to the current
+                        mountpoint, current workflow or the current workflow's data area. When selected a new drop-down menu
+                        appears to choose which of the three options to use.
+                    </li>
+                    <li><i>Custom/KNIME URL:</i> Allows to specify a URL (e.g. file://, http:// or knime:// protocol).
+                        When selected, a spinner appears that allows you to specify the desired connection and write timeout
+                        in milliseconds. In case it takes longer to connect to the host / write the file, the node fails to
+                        execute. Browsing is disabled for this option.
+                    </li>
+                </ul>
+                It is possible to use other file systems with this node. Therefore, you have to enable the file system
+                connection input port of this node by clicking the <b>...</b> in the bottom left corner of the node's icon
+                and choose <i>Add File System Connection port</i>.
+                <br />
+                Afterwards, you can simply connect the desired connector node to this node. The file system connection will
+                then be shown in the drop-down menu. It is greyed out if the file system is not connected in which case
+                you have to (re)execute the connector node first. Note: The default file systems listed above can't be
+                selected if a file system is provided via the input port.
+                <br/><br/>
+                <b>File/URL</b>
+                <br/>
+                Enter a URL when writing to <i>Custom/KNIME URL</i>, otherwise enter a path to a file. The
+                required syntax of a path depends on the chosen file system, such as &quot;C:\path\to\file&quot; (Local
+                File System on Windows) or  &quot;/path/to/file&quot; (Local
+                File System on Linux/MacOS and Mountpoint). For file systems connected via input port, the node
+                description of the respective connector node describes the required path format.
+                You can also choose a previously selected file from the drop-down list, or select a location
+                from the &quot;Browse...&quot; dialog.
+                Note that browsing is disabled in some cases:
+                <ul>
+                    <li><i>Custom/KNIME URL:</i> Browsing is always disabled.</li>
+                    <li><i>Mountpoint:</i> Browsing is disabled if the selected mountpoint isn't connected.
+                        Go to the KNIME Explorer and connect to the mountpoint to enable browsing.</li>
+                    <li><i>File systems provided via input port:</i>
+                        Browsing is disabled if the connector node hasn't been executed since the workflow has been opened.
+                        (Re)execute the connector node to enable browsing.</li>
+                </ul>
+                <i>The location can be exposed as or automatically set via a
+                </i><a href="https://docs.knime.com/latest/analytics_platform_file_handling_guide/index.html#path">
+                <i>path flow variable.</i></a>
+                <br/><br/>
+                <b>Create missing folders</b>
+                <br/>
+                Select if the folders of the selected output location should be created if
+                they do not already exist. If this option is unchecked, the node will fail if a folder does not exist.
+                <br/><br/>
+                <b>If exists</b>
+                <br/>
+                Specify the behavior of the node in case the output file already exists.
+                <ul>
+                    <li><i>Overwrite:</i> Will replace any existing file.
+                    </li>
+                    <li><i>Fail:</i> Will issue an error during the node's execution (to prevent unintentional overwrite).
+                    </li>
+                </ul>
+            </option>
+            <option name="Fingerprint column">The input column containing fingerprints (DenseBitVector cells).</option>
+            <option name="Id column">
+                The input column containing IDs that shall be written as second column into the FPS file.
+                It is possible to use Row IDs.
+            </option>
+           </tab>
+    </fullDescription>
+
+    <ports>
+        <inPort index="0" name="Input fingerprints">Table containing the fingerprints.</inPort>
+        <dynInPort insert-before="0"
+                   name="File system connection"
+                   group-identifier="File System Connection">
+            The file system connection.
+        </dynInPort>
+    </ports>
+</knimeNode>
